Renamed projects - web-backend -> Backend - web-frontent -> Frontend
