+++ Backend/WeatherForecast.cs
@@ -1,4 +1,4 @@
-namespace web_backend
+namespace Backend
 {
     public class WeatherForecast
     {
