+++ Frontend/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "test",
-  "version": "0.1.0.1",
+  "name": "Frontend",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@microsoft/signalr": "^6.0.8",
