+++ Frontend/src/components/LightBrightness.tsx
