+++ Frontend/src/components/LightColor.tsx
