+++ Frontend/src/components/LightWakeMode.tsx
