+++ Frontend/src/components/SmartLight.tsx
