+++ Frontend/src/models/DeviceConfig.tsx
