+++ Frontend/src/models/DeviceModel.tsx
