+++ Frontend/src/models/DeviceState.tsx
