+++ Frontend/src/models/PossibleDeviceTraits.tsx
