+++ Frontend/src/models/traits/BrightnessTrait.tsx
