+++ Frontend/src/models/traits/ColorTrait.tsx
