+++ Frontend/src/models/traits/WakeModeTrait.tsx
